+++ skoolkit/snaskool.py
@@ -108,7 +108,7 @@ class Entry:
         return self.block.get_ignoreua_directive(comment_type)
 
 class Disassembly:
-    def __init__(self, snapshot, ctl_parser, config=None, final=False, asm_hex=False, asm_lower=False):
+    def __init__(self, snapshot, ctl_parser, final=False, config=None, asm_hex=False, asm_lower=False):
         ctl_parser.apply_asm_data_directives(snapshot)
         self.config = config or {}
         dconfig = DisassemblerConfig(
@@ -239,7 +239,7 @@ class SkoolWriter:
     def __init__(self, snapshot, ctl_parser, options, config):
         self.comment_width = max(options.line_width - 2, MIN_COMMENT_WIDTH)
         self.asm_hex = options.base == 16
-        self.disassembly = Disassembly(snapshot, ctl_parser, config, True, self.asm_hex, options.case == 1)
+        self.disassembly = Disassembly(snapshot, ctl_parser, True, config, self.asm_hex, options.case == 1)
         self.address_fmt = get_address_format(self.asm_hex, options.case == 1)
         self.config = config